+++ CHANGELOG.rst
@@ -101,6 +101,9 @@ Recipe Changes
     This reflects their change in position within the game, from the primitive blast furnace to the
     Bessemer Converter.
 
+  - Hydrogen can now be made from a direct water gas shift recipe using refinery gas, without 
+    needing to distill it.
+
 - Tier 2 (MV) changes:
 
     - Vanadium steel is now made from a mixture of iron, ferrochrome, and ferrovanadium.